Import the latest changes from the development repository: - adjustments of the source code to compile with MS Visual C++ compiler; - add support for showing the validation progress in the Desktop Explorer UI; - refinements to testing and experiment-reproduction instructions; - parallelized saving of ledger snapshots; - automatic deletion of extraneous ledger snapshots; - update the copy of the parallelized validation paper; - correct reporting of peak memory usage on Mac OS; - atomic updates of statistics during parallel reward computations.

+++ lib/dt/index/block-fees.hpp
@@ -33,7 +33,7 @@ namespace daedalus_turbo::index::block_fees {
             blk.foreach_tx([&](const auto &tx) {
                 fees += tx.fee();
             });
-            idx.emplace_back(blk.issuer_hash(), fees, blk.offset() + blk.size(), blk.era());
+            idx.emplace_back(blk.issuer_hash(), fees, blk.offset() + blk.size(), static_cast<uint8_t>(blk.era()));
         }
     };
     using indexer = indexer_multi_epoch<item, chunk_indexer>;
